game logic is inside gamegrid now,  slight modifications

+++ src/snake/cell.rs
@@ -2,13 +2,13 @@ use std::io::Stdout;
 
 #[derive(Default)]
 pub struct Cell{
-    pub x: u32,
-    pub y: u32,
+    pub x: usize,
+    pub y: usize,
     pub object: Option<Box<dyn Display>>
 }
 
 impl Cell{
-    pub fn new(x: u32, y:u32) -> Cell{
+    pub fn new(x: usize, y:usize) -> Cell{
        Cell{
         x: x,
         y: y,
